Remove bogus pointer->integer cast
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@39011 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ ps/ps.map/ps_fclrtbl.c
@@ -214,7 +214,7 @@ int PS_fcolortable(void)
 	k = strlen(buf) - 1;
 	while (buf[k] == '0')
 	    k--;
-	k = k - (int)(strchr(buf, '.') - (unsigned)buf);
+	k = k - (int)(strchr(buf, '.') - buf);
 	ddig = k;
     }
